Added keybinds and gui to database for future use. Clean up code.
