+++ Projects/CoX/Data/sqlite/segs_game_sqlite_create.sql
@@ -7,10 +7,10 @@ CREATE TABLE `table_versions` (
     `last_update` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
-INSERT INTO table_versions VALUES(1,'db_version',3,'2018-01-28 10:27:01');
+INSERT INTO table_versions VALUES(1,'db_version',4,'2018-03-05 00:55:01');
 INSERT INTO table_versions VALUES(2,'table_versions',0,'2017-11-11 08:57:42');
 INSERT INTO table_versions VALUES(3,'accounts',0,'2017-11-11 08:57:43');
-INSERT INTO table_versions VALUES(4,'characters',4,'2018-01-28 10:16:27');
+INSERT INTO table_versions VALUES(4,'characters',5,'2018-03-05 00:54:27');
 INSERT INTO table_versions VALUES(5,'costume',0,'2017-11-11 08:57:43');
 INSERT INTO table_versions VALUES(6,'progress',0,'2017-11-11 08:57:43');
 INSERT INTO table_versions VALUES(7,'supergroups',0,'2018-01-23 10:16:43');
@@ -35,6 +35,8 @@ CREATE TABLE `characters` (
 	`endurance`	INTEGER DEFAULT 0,
 	`supergroup_id`	INTEGER NOT NULL DEFAULT 0,
 	`options`	BLOB,
+	`gui`		BLOB,
+	`keybinds`	BLOB,
 	FOREIGN KEY(`account_id`) REFERENCES accounts ( account_id ) ON DELETE CASCADE
 );
